Protocol (#154)
* send full arguments to client

* Add exception test

* Add test array type exception

* Clean up fmt

* Define branch if varaible undefined

* Improve virtualenv

* Rename device->class operation->function

* Add id

* Compute non explicit std::array length

* Clean up

* Start auto handling

* First draw. Require demangling.

* demangle type

* Handle arguments of auto returning functions

* Add test

* Test return type exception

* Start ExceptionTests

* Fix name

* Add exception tests

* Test connection exceptions

* Add test send uint32 vector

* Add tuple exception test

* Rm js api

* Remove python test

* Trig automation build

* Trig builds on current branch

* Use for loop

* Call corresponding branch to build

* Specify branchs by hand

* Fix variable name

* Fix script

* Test python from clone

* Try fix build

* Fix ci script

* Don't use script

* Try fix build

* Fix venv path

* Fix build

* Try fix build

* try fix build

* clean up

* Try fix venv cache

* Fix path

* Try fix venv cache

* show dir content

* Try fix venv cache

* clean up

* Simplify make.py

* Trig client tests only if test succeed

* Build clients only on master

* Clean up

* Add length before receive buffer

* Update send string

* refactor syslog

* Clean up

* Display output

* Fix parameter pack

* Create string function for parameter packs

* Move functions into string_utils

* Clean up

* Clean up

* priority as template parameter

* Use std::string

* Remove run-time if

* clean up

* Specify return types for server commands

* Rename kserver_syslog -> syslog

* Always ask the compiler to compute std::array length

* clean up

* Optimize send buffer copy

* clean up

* Add error message

* Add error message

* Improve error message

* Use signed instead of unsigned

* fix type

* Fix format

+++ core/websocket.hpp
@@ -29,7 +29,7 @@ class WebSocket
     template<class T> int send(const T *data, unsigned int len);
 
     char* get_payload_no_copy() {return payload;}
-    unsigned int payload_size() const {return header.payload_size;}
+    int64_t payload_size() const {return header.payload_size;}
     
     bool is_closed() const {return connection_closed;}
 
@@ -55,7 +55,7 @@ class WebSocket
     struct {
         unsigned int header_size;
         int mask_offset;
-        unsigned int payload_size;
+        int64_t payload_size;
         bool fin;
         bool masked;
         unsigned char opcode;
@@ -98,7 +98,7 @@ class WebSocket
     int read_stream();
     int read_header();
     int check_opcode(unsigned int opcode);
-    int read_n_bytes(int bytes, int expected);
+    int read_n_bytes(int64_t bytes, int64_t expected);
 
     int set_send_header(unsigned char *bits, long long data_len,
                         unsigned int format);
